[lldb][NativeRegisterContext] Rename to x86 for shared files
