+++ lldb/source/Plugins/Process/Linux/CMakeLists.txt
@@ -13,7 +13,7 @@ add_lldb_library(lldbPluginProcessLinux
   NativeRegisterContextLinux_ppc64le.cpp
   NativeRegisterContextLinux_riscv64.cpp
   NativeRegisterContextLinux_s390x.cpp
-  NativeRegisterContextLinux_x86_64.cpp
+  NativeRegisterContextLinux_x86.cpp
   NativeThreadLinux.cpp
   Perf.cpp
   Procfs.cpp
