+++ lldb/source/Plugins/Process/NetBSD/NativeRegisterContextNetBSD_x86.h
@@ -1,4 +1,4 @@
-//===-- NativeRegisterContextFreeBSD_x86_64.h -------------------*- C++ -*-===//
+//===-- NativeRegisterContextNetBSD_x86.h -----------------------*- C++ -*-===//
 //
 // Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
 // See https://llvm.org/LICENSE.txt for license information.
@@ -8,37 +8,35 @@
 
 #if defined(__i386__) || defined(__x86_64__)
 
-#ifndef lldb_NativeRegisterContextFreeBSD_x86_64_h
-#define lldb_NativeRegisterContextFreeBSD_x86_64_h
+#ifndef lldb_NativeRegisterContextNetBSD_x86_h
+#define lldb_NativeRegisterContextNetBSD_x86_h
 
 // clang-format off
 #include <sys/param.h>
-#include <sys/ptrace.h>
 #include <sys/types.h>
+#include <sys/ptrace.h>
 #include <machine/reg.h>
 // clang-format on
 
 #include <array>
 #include <optional>
 
-#include "Plugins/Process/FreeBSD/NativeRegisterContextFreeBSD.h"
-#include "Plugins/Process/Utility/NativeRegisterContextDBReg_x86.h"
+#include "Plugins/Process/NetBSD/NativeRegisterContextNetBSD.h"
 #include "Plugins/Process/Utility/RegisterContext_x86.h"
+#include "Plugins/Process/Utility/NativeRegisterContextDBReg_x86.h"
 #include "Plugins/Process/Utility/lldb-x86-register-enums.h"
 
-#define LLDB_INVALID_XSAVE_OFFSET UINT32_MAX
-
 namespace lldb_private {
-namespace process_freebsd {
+namespace process_netbsd {
 
-class NativeProcessFreeBSD;
+class NativeProcessNetBSD;
 
-class NativeRegisterContextFreeBSD_x86_64
-    : public NativeRegisterContextFreeBSD,
+class NativeRegisterContextNetBSD_x86
+    : public NativeRegisterContextNetBSD,
       public NativeRegisterContextDBReg_x86 {
 public:
-  NativeRegisterContextFreeBSD_x86_64(const ArchSpec &target_arch,
-                                      NativeThreadFreeBSD &native_thread);
+  NativeRegisterContextNetBSD_x86(const ArchSpec &target_arch,
+                                     NativeThreadProtocol &native_thread);
   uint32_t GetRegisterSetCount() const override;
 
   const RegisterSet *GetRegisterSet(uint32_t set_index) const override;
@@ -54,7 +52,7 @@ public:
   Status WriteAllRegisterValues(const lldb::DataBufferSP &data_sp) override;
 
   llvm::Error
-  CopyHardwareWatchpointsFrom(NativeRegisterContextFreeBSD &source) override;
+  CopyHardwareWatchpointsFrom(NativeRegisterContextNetBSD &source) override;
 
 private:
   // Private member types.
@@ -62,6 +60,7 @@ private:
     GPRegSet,
     FPRegSet,
     DBRegSet,
+    MaxRegularRegSet = DBRegSet,
     YMMRegSet,
     MPXRegSet,
     MaxRegSet = MPXRegSet,
@@ -69,11 +68,9 @@ private:
 
   // Private member variables.
   std::array<uint8_t, sizeof(struct reg)> m_gpr;
-  std::array<uint8_t, 512> m_fpr; // FXSAVE
+  std::array<uint8_t, sizeof(struct xstate)> m_xstate;
   std::array<uint8_t, sizeof(struct dbreg)> m_dbr;
-  std::vector<uint8_t> m_xsave;
-  std::array<uint32_t, MaxRegSet + 1> m_xsave_offsets;
-  std::array<size_t, MaxRegSet + 1> m_regset_offsets;
+  std::array<size_t, MaxRegularRegSet + 1> m_regset_offsets;
 
   std::optional<RegSetKind> GetSetForNativeRegNum(uint32_t reg_num) const;
 
@@ -89,9 +86,9 @@ private:
   std::optional<YMMSplitPtr> GetYMMSplitReg(uint32_t reg);
 };
 
-} // namespace process_freebsd
+} // namespace process_netbsd
 } // namespace lldb_private
 
-#endif // #ifndef lldb_NativeRegisterContextFreeBSD_x86_64_h
+#endif // #ifndef lldb_NativeRegisterContextNetBSD_x86_h
 
-#endif // defined(__x86_64__)
+#endif // defined(__i386__) || defined(__x86_64__)
